Adding additional arcade filters
These new arcade filters are meant to compliment the existing Arcade one. They represent what you'd might see playing a game with monitor display tech from the early 80s and the 90s. Updated the existing Arcade.ini (and vertical) so it's description is similar to the others and identifies it as a late 80s preset.